[#2342] replace size_t

+++ nes-core/include/Operators/LogicalOperators/Sources/MaterializedViewSourceDescriptor.hpp
@@ -30,7 +30,7 @@ class MaterializedViewSourceDescriptor : public SourceDescriptor {
      * @brief Factory method ***
      */
     static SourceDescriptorPtr create(const SchemaPtr& schema,
-                                      size_t mViewId);
+                                      uint64_t mViewId);
 
     /**
      * @brief Provides the string representation of the materialized view source
@@ -49,15 +49,15 @@ class MaterializedViewSourceDescriptor : public SourceDescriptor {
     * @brief returns the materialized view id
     * @return
     */
-    size_t getViewId() const;
+    uint64_t getViewId() const;
 
   private:
     /**
       * @brief **
       */
-    explicit MaterializedViewSourceDescriptor(SchemaPtr schema, size_t viewId);
+    explicit MaterializedViewSourceDescriptor(SchemaPtr schema, uint64_t viewId);
     SchemaPtr schema;
-    size_t viewId;
+    uint64_t viewId;
 };
 using MaterializedViewSourceDescriptorPtr = std::shared_ptr<MaterializedViewSourceDescriptor>;
 }// namespace NES::Experimental::MaterializedView
