+++ nes-core/include/Sinks/SinkCreator.hpp
@@ -216,7 +216,7 @@ DataSinkPtr createNetworkSink(const SchemaPtr& schema,
 DataSinkPtr createMaterializedViewSink(SchemaPtr schema,
                                        const Runtime::NodeEnginePtr& nodeEngine,
                                        QuerySubPlanId parentPlanId,
-                                       size_t viewId);
+                                       uint64_t viewId);
 
 #ifdef ENABLE_KAFKA_BUILD
 /**
